refact: optimizing and refactoring wallet performance and coincap implementation

+++ src/main/java/com/example/cryptowalletmanagement/dto/wallet/WalletEvaluationInputDTO.java
@@ -1,7 +1,8 @@
 package com.example.cryptowalletmanagement.dto.wallet;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public record WalletEvaluationInputDTO(List<AssetInputDTO> assets) {
-    public record AssetInputDTO(String symbol,Double quantity,Double value) {}
+    public record AssetInputDTO(String symbol, BigDecimal quantity, BigDecimal value) {}
 }
